Add basic Login and Sign Up functionality
Now the client can connect to the server and send login/sign up requests. It shows the result of the request in the UI as well.

At this point, however, certain things are not handled (mainly by server). For example, any password is allowed when signing up (even empty passwords!). Also accounts do not have profile pictures.

Also added run configurations for IDEA.