Add timeline functionality to server
Now the server can show a user's "timeline", which consists of the user's own posts and posts from those the user follows.

Also for the results to be sorted by time, PostDao had to be changed a bit so it always sorts the result.

Also added test codes and test configurations.