Add missing timeline features to client
Now the timeline shows profile photos of users whose posts are in the timeline.

Also profile picture of current logged in user is shown at the top of the window.

Whenever a photo cannot be shown due to an error, an icon with an exclamation mark is shown (error.png and error_small.png).

Also the looks of the views has been revised to give a more proper look to the client app.

For more convenience, Utils.java is added that contains useful methods for loading user info and loading images from server.